sml doc comments added; href, target and title attribute mappers created;
