+++ sml/attributes.go
@@ -2,8 +2,6 @@ package sml
 
 import (
 	"fmt"
-	"html"
-	"net/url"
 	"strings"
 
 	"github.com/Drolfothesgnir/shitposter/scum"
@@ -36,49 +34,4 @@ func handleAttributes(b *strings.Builder, w *[]string, m attrMap, n scum.Seriali
 		b.WriteByte(' ')
 		b.WriteString(attr.String())
 	}
-}
-
-func attrHref(b *strings.Builder, w *[]string, a scum.SerializableAttribute) bool {
-	if a.IsFlag {
-		*w = append(*w, "attribute href must have a value")
-		return false
-	}
-
-	payload := strings.TrimSpace(a.Payload)
-	if payload == "" {
-		*w = append(*w, "attribute href must not be empty")
-		return false
-	}
-
-	if strings.ContainsAny(payload, "\x00\r\n\t") {
-		*w = append(*w, "attribute href contains forbidden control characters")
-		return false
-	}
-
-	u, err := url.Parse(payload)
-	if err != nil {
-		*w = append(*w, fmt.Sprintf("attribute href is invalid: %v", err))
-		return false
-	}
-
-	switch strings.ToLower(u.Scheme) {
-	case "":
-		// Allow relative references, but reject protocol-relative URLs such as //evil.com.
-		if strings.HasPrefix(payload, "//") {
-			*w = append(*w, "attribute href must not be protocol-relative")
-			return false
-		}
-
-	case "http", "https", "mailto":
-		// Allowed schemes.
-
-	default:
-		*w = append(*w, fmt.Sprintf("attribute href scheme %q is not allowed", u.Scheme))
-		return false
-	}
-
-	b.WriteString(`href="`)
-	b.WriteString(html.EscapeString(payload))
-	b.WriteByte('"')
-	return true
 }
